+++ src/main/java/com/coffeecart/ui/elements/TotalButtonElement.java
@@ -1,9 +1,10 @@
-package com.coffeecart.ui.component;
+package com.coffeecart.ui.elements;
 
+import com.coffeecart.ui.component.BaseComponent;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 
-public class TotalButtonElement extends BaseComponent{
+public class TotalButtonElement extends BaseComponent {
     public TotalButtonElement(WebDriver driver, WebElement rootElement) {
         super(driver, rootElement);
     }
