+++ src/main/java/com/coffeecart/ui/modal/SuccessfulPopUp.java
@@ -1,9 +1,10 @@
-package com.coffeecart.ui.elements;
+package com.coffeecart.ui.modal;
 
+import com.coffeecart.ui.elements.BaseElement;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 
-public class SuccessfulPopUp extends BaseElement{
+public class SuccessfulPopUp extends BaseElement {
     public SuccessfulPopUp(WebDriver driver, WebElement rootElement) {
         super(driver, rootElement);
     }
